alice/pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-alice"
version = "0.1.0"
description = "Alice environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "RestrictedPython>=7.0",
    "openai>=1.0",
    "numpy>=1.24",
    "requests>=2.28",
]

[project.optional-dependencies]
dev = [
    "hypothesis>=6.0",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m alice.server.app
server = "alice.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["alice", "alice.server"]
package-dir = { "alice" = ".", "alice.server" = "server" }